add print

+++ db/compaction/compaction_job.cc
@@ -804,7 +804,7 @@ Status CompactionJob::Install(const MutableCFOptions& mutable_cf_options) {
 
   // zhangxin
   int64_t current_time = env_->NowMicros();
-  int max_live_time0 = 0, max_live_time1 = 0;
+  int64_t max_live_time0 = 0, max_live_time1 = 0;
   if (env_->GetCurrentTime(&current_time).ok()) {
     std::string sst_live_str;
     for (int level = 0; level < 2 && level < vstorage->num_levels(); level++) {
@@ -819,9 +819,9 @@ Status CompactionJob::Install(const MutableCFOptions& mutable_cf_options) {
           sst_live_str.append(buf);
 
           if (level == 0) {
-            max_live_time0 = std::max(max_live_time0, current_time-creation_time);
+            max_live_time0 = std::max(max_live_time0, int64_t(current_time-creation_time));
           } else if (level == 1) {
-            max_live_time1 = std::max(max_live_time1, current_time-creation_time);
+            max_live_time1 = std::max(max_live_time1, int64_t(current_time-creation_time));
           }
         }
       }
